Decouple boot table from S29JL driver
