+++ libs/drivers/CMakeLists.txt
@@ -16,3 +16,4 @@ add_subdirectory(watchdog)
 add_subdirectory(uart)
 add_subdirectory(program_flash)
 add_subdirectory(gyro)
+add_subdirectory(s29jl)
