+++ libs/drivers/program_flash/Include/program_flash/boot_table.hpp
@@ -4,6 +4,7 @@
 #include <cstdint>
 #include <gsl/span>
 #include <tuple>
+#include "base/os.h"
 #include "flash_driver.hpp"
 #include "utils.h"
