+++ libs/drivers/s29jl/CMakeLists.txt
@@ -1,0 +1,15 @@
+set(NAME flash_s29jl)
+
+set(SOURCES
+    s29jl.cpp
+)
+
+add_library(${NAME} STATIC ${SOURCES})
+
+target_link_libraries(${NAME} 
+	base
+	program_flash
+)
+
+target_include_directories(${NAME} INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/Include)
+target_include_directories(${NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/Include/flash)
