+++ libs/obc/hardware/CMakeLists.txt
@@ -25,6 +25,7 @@ target_link_libraries(${NAME}
 	rtc
 	uart
 	comm
+	flash_s29jl
 )
 
 target_include_directories(${NAME} INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/Include)
